Base pinned site implementation
I'd like to get this landed and work on edges in follow ups.  Persistent state isn't implemented yet so pinned tabs won't be that useful until that is done.  Drag and drop for creating pinned sites and unpinning via drag and drop is not implemented yet.

Reviewer: @diracdeltas
