+++ less/tabs.less
@@ -28,7 +28,6 @@
 }
 
 .tab {
-  background: linear-gradient(to bottom, #BEBFC2, #AFB0B4);
   border-radius: 9px;
   border-top-left-radius: @tabBorderRadius;
   border-top-right-radius: @tabBorderRadius;
@@ -46,6 +45,10 @@
   width: 100%;
   -moz-window-dragging: no-drag;
 
+  &:not(.isPinned) {
+    background: linear-gradient(to bottom, #BEBFC2, #AFB0B4);
+  }
+
   .tabTitle {
     -webkit-user-select: none;
     box-sizing: border-box;
@@ -62,7 +65,7 @@
     background-position: center;
     background-repeat: no-repeat;
     display: inline-block;
-    margin: auto 4px auto 9px;
+    margin: auto 9px;
     position: relative;
     vertical-align: middle;
   }
@@ -144,12 +147,15 @@
   display: inline-block;
   padding: 0 2px 0 0;
   position: relative;
-  flex: 1;
   vertical-align: top;
   // There's a special case that tabs should span the full width
   // if there are a full set of them.
   &:not(.partOfFullPageSet) {
     max-width: 184px;
+  }
+
+  &:not(.isPinned) {
+    flex: 1;
   }
 
   hr.dragIndicator {
@@ -176,6 +182,7 @@
 }
 
 .tabsToolbar {
+  background: @tabsBackground;
   display: flex;
   height: 26px;
 }
@@ -250,4 +257,8 @@
   padding-right: 10px;
   text-align: right;
   transform: translateX(-20px);
+}
+
+.pinnedTabs {
+  margin-left: 9px;
 }
